Merged PR 4757835: Enabling FactoryOrchestrator External Tasks to Support Instructional Image and Video
The code change is entirely app-side. The user specifies the path of the external task to be a file path on the system service side. When an ExternalTest is navigated to on the ExternalTestResultPage, the code will attempt to discern information about the file path, like if it is an image or video. The code does this by only supporting some types of image and video file extensions. If the file extension of the incoming media matches one of the supported image or video types, the code will then handle the file using the appropriate mechanisms, and either display an image or video on the UI.

Additionally, the UI for ExternalTestResultPage is remodeled.

- Implemented image and video support on external tasks
- Added visual cues for image and video not found exceptions, and remodeled UI for external task
- updated version to be 4.1

Related work items: #26510266